rename: Rename CalendarModal.tsx due to case issue
